Enable setting the office hours cta from Layout component
Also adjusts footer styles based on the presence of the CTA or not.
