+++ src/layout/footer/style.css
@@ -1,4 +1,4 @@
-.pds-site-footer {
+.pds-site-footer:not(.with-cta) {
   border-top: 1px solid var(--pds-color-border-default);
 }
