+++ src/layout/layout/index.js
@@ -4,11 +4,22 @@ import Footer from "../footer"
 
 import SearchBar from "../SearchBar"
 
-import { Container } from "@pantheon-systems/pds-toolkit-react"
+import { Container, CTASlice } from "@pantheon-systems/pds-toolkit-react"
 
 import "./style.css"
 
-const Layout = ({ children, containerWidth, pageType = "default" }) => {
+// Content for CTA
+const primaryCTA = {
+  text: "Learn Pantheon",
+  url: "https://pantheon.io/learn-pantheon?docs",
+}
+
+const secondaryCTA = {
+  text: "Office Hours",
+  url: "https://pantheon.io/developers/office-hours?docs",
+}
+
+const Layout = ({ children, containerWidth, hasCta, pageType = "default" }) => {
   return (
     <div className="pantheon-docs">
       <Header page={pageType} />
@@ -16,7 +27,16 @@ const Layout = ({ children, containerWidth, pageType = "default" }) => {
         <SearchBar page={pageType} />
       </Container>
       {children}
-      <Footer />
+      {hasCta && (
+        <CTASlice
+          backgroundColor="secondary"
+          headingText="Got questions? We've got answers!"
+          primaryLink={primaryCTA}
+          secondaryLink={secondaryCTA}
+          className="pre-footer-slice"
+        />
+      )}
+      <Footer className={hasCta ? "with-cta" : null} />
     </div>
   )
 }
